+++ AvgScript/src/extensions/_include.ts
@@ -1,0 +1,3 @@
+import './arrayEx';
+import './mapEx';
+import './stringEx';